requirements.txt
boto3
pandas
fsspec
s3fs
aws-cdk-lib==2.26.0
constructs>=10.0.0,<11.0.0
amazon-textract-idp-cdk-constructs>=0.0.5
schadem-tidp-manifest==0.0.8
